Fix BrotliFillBitWindow (non-compile-time n_bits version)
When IS_CONSTANT(n_bits) is false, most of the if statements will never
pass in the original C version, so remove them.

I didn't measure any significant performance impact.

(The code is also slightly incorrect: the 'BROTLI_ALIGNED_READ == 0' is
in the wrong if statement in the first two lines removed by this patch.)