Remove `ideal_width` rustfmt config
**This Commit**
Updates the root `rustfmt.toml` to no longer contain the
`ideal_width` configuration.

**Why?**
It was renamed to `comment_width` in [this PR][pr] and its default value
is 80 so it didn't seem valuable to specify.

**Notes**
There are other values specified to be their defaults in the
`rustfmt.toml` - `reorder_imports` and `max_width`. We could remove
these too and just leave the override for `tab_spaces`. We could also
delete the file and just use the defaults/standards for everything but I
figured I'd propose one thing at a time.

[pr]: https://github.com/rust-lang/rustfmt/pull/1370