Auto-fixed clippy::identity_op
Note that this one might point to the underlying bug in the original implementation -- the masking of a `u8` values happens after casting it to a `u8`.

This was automatic change using these commands, and later excluded one that does not look that good.

```sh
cargo clippy --fix -- -A clippy::all -W clippy::identity_op
cargo fmt --all
```

* https://rust-lang.github.io/rust-clippy/master/index.html#/identity_op