Simplify constant func params
Whenever calling a function on a value, e.g. `value.wrapping_sub(1usize)`, there is almost never a need for specifying constant type.  The theoretical exception is if calling `foo<T>(&self, value: T)`, but I don't think there is a case like this here.

Another common case is a constant with type as an indexer - also never needed unless there is some magical implementation of the indexer trait.

Executed replacements:

```rust
.foo(123_usize)    -> .foo(123)
.foo(123 as usize) -> .foo(123)
[123_usize]        -> [123]
[(123_usize)]      -> [123]
[(123)]            -> [123]
[123_usize..       -> [123..
[(123_usize)..     -> [123..
```

```regex
(\.[a-z_]+)\((\d+)_?[a-z][a-z0-9]*\)    ->  $1($2)
(\.[a-z_]+)\((\d+) as [a-z][a-z0-9]*\)  -> $1($2)
\[(\d+)_?[a-z][a-z0-9]*\]               -> [$1]
\[\((\d+)_?[a-z][a-z0-9]*\)\]           -> [$1]
\[\((\d+)(_?[a-z][a-z0-9]*)?\)\]        -> [$1]
\[(\d+)_?[a-z][a-z0-9]*\.\.             ->  [$1..
\[\((\d+)_?[a-z][a-z0-9]*\)\.\.         ->  [$1..
```