Fix typed consts as params
Whenever a number is used as a parameter to a function, i.e. has a comma or an opening `(` before it, and if the number has a type, remove that type.

`foo(0u8, 100i32, "bar")` -> `foo(0, 100, "bar")`

Note that this does not address `(10u32).saturated_add(1)` -- which is on purpose.

Note that this also fixes hard-coded arrays, but does not remove the type of the first element:  `[0u8, 1u8, 2u8]` -> `[0u8, 1, 2]` -- which keeps the array type as before.

Run this twice, and use `cargo fmt`.  ONLY DO IT ON *.rs files.

```regex
([(,]\s*)(-?\d+)(_?|\s*as\s*)[ui][a-z0-9]*(\s*,)        -> $1$2$4
(,\s*)(-?\d+)(_?|\s*as\s*)[ui][a-z0-9]*(\s*[)\],])      -> $1$2$4
(,\s*)\((-?\d+)(_?|\s*as\s*)[ui][a-z0-9]*\)(\s*[)\],])  -> $1$2$4
```