Use proper `self` instead of `xself`
There are a lot of hacky `xself` that should use `self`.  This PR currently has only one minor change, but I would like to change all/most of them in the similar manner. Note that the usage here is a bit strange - a struct is created only to be discarded right after.

Are there any reasons to preserve `xself`?