Remove unused corrosion code
The corrosion dir and `uncorrode.py` seem to no longer be needed.

The code is still available in git history, so it is not lost.

Removing it here allows contributors to focus on just the relevant code without confusion, and prevents "surprising hidden code" attacks like in the xz case.