Use bool instead of i32
For public API, add a deprecated attribute, but the replacement function is NOT being made public just yet.  In the future, we should review and drastically reduce the public API - this crate should only expose what's needed, rather than expose everything and be responsible to keep the internal structure intact.

Additionally, reworked `ShouldCompress` and UTF8 parsing