Use bools instead of i32s (second part)
There are a few publicly visible functions that should also be using a `bool` type.  In order not to cause an API change, I created a crate-visible variants so that we can remove them in the next major release. In the meantime, users will see a deprecation warning IF they are using these functions (most likely they are not) - but if they are, they will let us know to keep them public.

Note a few extra `FIXME` -- these I think we should fix when making a breaking change, but not just yet because they are harder to do in a backward-compatible way