Adjust license to reflect a MIT-only source, and add a MIT license file
Since `entropy_encode.rs` appears to be derived from
https://github.com/google/brotli/blob/v1.1.0/c/enc/entropy_encode.c
(possibly at a different version), and carries only MIT terms, it’s not
possible for the crate to be `BSD-3-Clause OR MIT` overall. The new
`LICENSE.MIT` file is taken from
https://github.com/google/brotli/raw/refs/tags/v1.1.0/LICENSE since that
appears to be the correct license text for `entropy_encode.rs`.

Fixes #41.

Signed-off-by: Benjamin A. Beasley <code@musicinmybrain.net>